+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.CriminalIntentApp">
         <activity
-            android:name=".presentation.RegisterActivity"
+            android:name=".presentation.authentication.RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.LoginActivity"
+            android:name=".presentation.authentication.LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
