+++ app/src/main/java/com/example/criminalintentapp/presentation/fragments/crime_list/CrimeListFragment.kt
@@ -16,7 +16,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.criminalintentapp.R
 import com.example.criminalintentapp.data.database.Crime
-import com.example.criminalintentapp.presentation.LoginActivity
+import com.example.criminalintentapp.presentation.authentication.LoginActivity
 import com.google.firebase.auth.FirebaseAuth
 
 class CrimeListFragment : Fragment(R.layout.fragment_crime_list) {
