+++ app/src/main/res/layout/activity_login.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.LoginActivity">
+    tools:context=".presentation.authentication.LoginActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
